Combine "sign in link" translation strings
Delete extraneous sign-in.promo-1 (which is already translated in the
share.sign-in string) and add a concatenated share.sign-in-2 string.

Once this is merged, I will bring over the translations form
sing-in.link and sign-in.promo-2 to the new share.sign-in-2. Then I’ll
make a new PR to delete the unused sign-in strings.