Make Chinese (Traditional) & Portuguese (Br) level 2
Chinese and Portuguese translations are moving along and with that comes the need to view the translations on Staging. Currently, they are only accessible via the level 1 switch. However, level 1 has so many languages that Chinese gets cut off (and Japanese is entirely off the screen). In order to allow translators to see their work, we need to bump these up to level 2 so they can see translations from there.