Update "Park(ing) Day" link
The http://parkingday.org/ website no longer works (the domain has expired). There is a Facebook group (https://www.facebook.com/PARKing-Day-71298180414\) that has a post from the most recent Parking Day in 2017, so it seems active. I suggest we use that link instead- it is the closest thing I could find to a central page for the event and should be relatively permanent as it's hosted on Facebook.