Add Simplified Chinese + related updates
Our Chinese translation has evolved since it was originally created. We
now will be launching a Traditional and Simplified Chinese translation
of Streetmix.

This PR adds Simplified as a language option and adjusts labels to fit
having two versions. A Chinese translator (Crystal) was consulted on
these labels, and recommended not using parentheses for our context. It
also updates the new translation strings for this in main.json.

The PR also re-orders the list in locales.json to be by 2 letter
language code (this is just to set up a way to keep the file organized
moving forward; it does not affect how the languages are displayed in
the UI). Currently, it looks like languages are being automatically
sorted in the UI by Unicode character codes in the native language's
spelling (e.g. Suomi (Finnish) goes after Polish (Polski), and 日本語
(Japanese) comes next as its script comes afterwards in Unicode). This
is a good way of sorting, and it makes sense for us to keep this way
for now. Note: there are multiple approaches to sorting languages. See
https://ux.stackexchange.com/questions/17064/how-do-you-sort-a-list-of-l
anguages and
https://en.wikipedia.org/wiki/Wikipedia:Language_order_poll#General_orde
r for perspectives on this.

Russian has also been added to assets/scripts/locales/locale.js, since
we forgot to do that in #1085.