fix(receiveStreet): do not call after catching error
A .then(receiveStreet) will always be called because it is placed after the
.catch() in the promise chain. Because the errorReceiveStreet() handler
does not return data (there is no data to pass along), receiveStreet() will
be called without transmission data. This is logged in Sentry:

https://sentry.io/share/issue/e226fe552dbc417786955ed22e571346/

The API is designed to _always_ return a transmission when successful,
so it's best to adhere to the expectations of that API. This is why
receiveStreet() is not modified to test that `transmission` is defined.
It should always be defined. In practice when it is not defined, it is
because it was called improperly.

There was a previous commit (7bbc95c40975bbaa0c68950da18de8f7b46d7c3b)
which moved .then(receiveStreet) after the .catch() because it seemed
that the error was being swallowed. However, this is likely because the
error was not being thrown properly in the first place, a fix that was
later introduced (31d8e31d3a98f1d3ef4d6fc9063a1343c76dec79). With this
in place, the reverting order of .then() and .catch() should behave as
expected.