fix(drag and drop): handle when oldDraggingState is undefined
A previous fix (see issue #1842) did not go far enough in fixing the problem.
`oldDraggingState` can also be undefined. This is tracked in Sentry:

https://sentry.io/share/issue/89df52be1b1e4e20a604561b5d1871e2/

I've updated the code to not just conservatively test equality for `null`
but on any falsy value, which would also cover undefined. This should be fine;
`oldDraggingState` is an object, and should never be a valid falsy value
we need to work with (e.g. `0` or `false`).