refactor(Avatar): convert to <span> element
When Avatar's root element is a <div> we receive a warning that it cannot be
used as the child of a <p> element. Since <Avatar> is designed to be an
inline-block (a pure inline <span> would not allow for absolute placement of
images within it) we may as well use the <span> element instead so that it
is semantically more in line with actual usage.