refactor(authentication): remove remaining code related to auth headers
Since JSON Web Tokens are passed via cookies, and an Authorization header is no longer
used, we remove all remaining code related to this Authorization header. Middleware used
to parse the header is now removed and will no longer be performed on any request.
The header is removed from tests and documentation. Finally, fix a bug in /services/images/
where the userId was formerly read from the legacy authentication data.

This leaves some cleanup work for the future:

1. If we decide to keep the Apiary docs (we might not; it's outdated), we should make
clear that authenticated requests are now passed through with cookies.
2. Do we mock authenticated requests for tests for the API, and how?