refactor(buttons): make primary, secondary and tertiary styles official
For a long time, button UI was in a transition state between an older style (borders on top
and bottom, squared corners) to a new style (no borders, round corners, brighter colors
matching the new UI color scheme). Applying the new style to buttons all at once was difficult
because there are other types of buttons (e.g. menu bar items, the close button) that shouldn't
be targeted for change. As a result, buttons in more-recently built / refactored components
got the new button styles first. (This is why the sign in dialog and many dialog boxes had
a green button, even though in other parts of the UI the button remained blue.)

This PR finally standardizes all "default" buttons to use the same consolidated CSS.

By default, buttons will a brighter shade of light blue than the legacy version.
By applying the class `button-primary`, you can get a call-to-action style green button.
By applying the class `button-secondary`, you get a darker blue button, like the sign in button.
By applying the class `button-tertiary`, you get a lower-priority button that only has a border
and a transparent background (the social sign-on options use this).

This unification of styles will make it easier to modify all buttons at once when the UI changes.

For example, all buttons now have a standardized hover state (it darkens slightly). Previously
when buttons had to be updated individually, this behavior was inconsistent or nonexistent.

One other change that happens here is that the infobubble variant buttons have changed to be part
of this system as well. Now, seelected variants will have a highlighted copper background color
instead of being grayed out.