fix(StreetNameplateContainer): regression in editable hover state
A previous refactor to convert `<StreetNameplateContainer />` to a function component
changed a prop name passed to the child component `<StreetName />` from `editable` to
`isEditable`. This caused `<StreetName />`'s `editable` value to pick up the default
value of `false`, causing it to not properly render the `hover` state.

The tests for editable hover state have been moved from the `<StreetName />` unit test
suite to the `<StreetNameplateContainer />` level to ensure that this integration between
components does not break. A small unit test remains in `<StreetName />` to ensure that
the default editable state is always false (e.g. when the streetname is displayed in
the gallery or in in-line text)