fix(sentiment survey): override disabled button style
Fixes a bug where the button style took on the global disabled button
state, with a light gray background. It now displays as transparent
as intended.

Note that we are also standardizing to the :disabled pseudo-selector
here over the [disabled] attribute selector. The pseudo-selector works
in all browsers that we support. There's an interesting discussion here
about how :disabled decouples the state from DOM.

https://stackoverflow.com/questions/20141450/should-i-use-css-disabled-pseudo-class-or-disabled-attribute-selector-or-is-i

As a follow up task we should replace [disabled] elsewhere but this
is out of scope for this commit.