feat(auth): simplify flow for signing in again
In the event that a user's client-side authentication state is incorrect (or outdated,
expired, corrupted, etc), we must effectively reset their client-side state and ask
them to sign in again. This can happen when we refactor or update our authentication
process, and the expected sign in state has changed. This is part of an ongoing process
to catch these conflicts, which have locked users out from using Streetmix.

In the past, when an authentication error has occurred, we display a "blocking error"
message and ask the user to reload the page. Pressing the reload button will then
clear authentication data, and force the window to reload. However, the blind spot here
is that clearing data happens _after_ the button is clicked, which means the usual
ways for reloading a page in a browser (e.g. pressing Ctrl-R) will not actually clear
that data, as expected.

This commit does not fix this problem universally, but does improve the process for the
AUTH_EXPIRED case. When we know this error is displayed, automatically clear all
old authentication data -- don't wait for user interaction to do it by pressing a
specific button. Once that has happened, we no longer need to force the page reload,
only to have the user find the sign-in button again. By automatically clearing
old authentication data, the button that can be displayed is the "Sign in" button
directly.

In other words, we reduce the "sign in again" friction from two screens / two buttons
down to one.

Unfortunately, dialog boxes are not designed to be displayed _above_ the blocking
error message. This commit also introduces a layout change at the app level so that
dialog boxes can be displayed on top of the error screen, so  that the Sign in
dialog works and is shown as intended.