refactor: path definitions include initial slash
This change makes strings for absolute paths more clear by including the initial slash (`/`)
in the string. This also avoids situations where we must manually strip the initial slash
from global variables (e.g. `window.location.pathname`) that include it, and or have to add
the initial slash when routing the browser.

This commit also changes how "url parts" are compared during page load. In theory, a direct
string equality comparison achieves the same thing as splitting the path on the slashes and
comparing individual string fragments, for most paths. The only time we need to split the
path is when we have paths such as `user/1/street-name`.