refactor(Palette): clean up layout styling
This is a pretty large refactor to the Palette component to clean up prototyping messiness
with the environment editor button, and to take advantage of flexbox layout (which is well
supported now, and many other UI components have been ported to flexbox over the years).

- command areas are now present on both sides of the palette bar, and layout is now using
  flexbox.
- duplicate styling for the left-side command area is removed.
- removed unused? z-index that only applied to the command buttons for some reason
- <Palette> is now renamed to <PaletteItems> for clarity, as it is a child of
  <PaletteContainer>
- command buttons are at a fixed height instead of being calculated based on palette bar
  height
- simplify height variables and calculation so that numbers look less "magic"

Overall, this cleans up the logic and should make it easier to modify code to work
within the Palette in the future.