fix(environs): disable supermoon, enable night
`supermoon` was enabled over `night` in a one-off situation when the Super Blood Wolf Moon was
a thing. It was then modified to have a normal moon. (At the time, `night` did not have a
moon.)

Now that the moon is just a normal, accepted part of the night view, restore to the original
`night` setting, add the normal moon to it, and reserve the `supermoon` setting for the future.