fix(GeotagDialog): do not stretch out error banner across the dialog box
The error banner on the GeotagDialog component was modeled after the NotificationBar
component, which stretches across the entire viewport. The error banner stretched
across the entire width of the dialog box. This meant that the "close" button on the
dialog box was placed _on_ the error banner itself, and that caused confusion about
what that close button did. For comparison, the close button on the original
NotificationBar component dismissed the bar itself, not the entire app. So it is
reasonable for a user to suspect that the close button on the GeotagDialog error
banner dismissed the erorr banner, not the dialog box.

By making the error banner not the full width of the dialog, the dialog "close"
button is now outside of the banner, which helps resolve that ambiguity.
The error banner still floats above the map as an important notice for users.