fix(UpDownInput): text input should update proper segment
There is a bug in the <UpDownInput /> component where updating its value
in the input element updates the wrong segment. This is caused by the
`debounceUpdateValue` method caching callback functions set to the position
of segments from previous renders.

This becomes solved by converting the component to a function component
(using React Hooks), from a class component.

In addition, the logic has been refactored to try and improve how the
component reacts / behaves. The next step is to follow up with better
unit tests.