fix(UpDownInput): work around user input bugs
We're trying to work around an issue (see #1766) where manual user input could be
prematurely "cancelled" or affected in unexpected ways. I suspect the issue can be
better solved by having application-level understanding of "dirty" input versus
"committed" input, but that's not present right now. We can also work around this
issue by removing the auto-update functionality from this component.

Rather than remove it completely, I've now placed it behind a `allowAutoUpdate` prop,
which is `false` by default. I'm allowing the building input to continue using this,
since its input mechanics are not quite as complex as the width input. If we figure
how to fix the root issue we can look into re-enabling this by default.

This now means width input is generally not committed until a user presses "Enter"
or blurs away from the input. Currently a mouse-out will commit the input as well.
This might be further tweaked in the future.

Other bug fixes:
- If a user types an input, then uses the +/- buttons, interacting with the input
  again should not restore the previous dirty input.
- Include a handler to manually re-implement double-clicking on input text to
  select it. This probably doesn't need to be done if we remove some of the existing
  "clever" UI behavior.