refactor(analytics): move capacity logic module
the 'util' folder should be for pretty generic utility functions.

capacity calculation is more specific functionality and it should live with capacity.json.
should capacity be in "street" or "segment"? you could make an argument for either one, but
since capacity.json is already there and capacity is largely defined per-segment let's leave
it in "segments" directory.