Postgres setup from the command line should read from env vars (#2162)
* chore(env): comment out optional env vars in env.example

In certain environments, env vars that are equal to nothing are not
treated as undefined. They are treated as being equal to the empty
string. This small tweak leaves optional env vars commented out so
if someone duplicates `.env.example` to `.env` as is, and makes no
changes to those lines, no new unexpected behavior could be
introduced by the sudden presence of env vars being equal to empty
strings.

* fix(postgres): config script should read env vars