refactor(menus): only mount menus to DOM when active
This change brought about because menus are mounted but set to 0 opacity
initially, so they're technically always present in the app. One side
effect of this is that the Facebook Container extension from Mozilla
will aggressively put a "Facebook is blocked" icon floating in the
application wherever the share menu happens to be in space when it starts.

One potential fix would be to conditionally render the Facebook share
link based on whether the menu is active, but not rendering menus until
they're active seems like the more long-term supportable solution.
This is also how dialogs work, and in general, the strategy we use for
most other components in the app.