chore(packages): remove postcss from package.json
Recently, postcss was bumped to v8, and this revealed that the dependency
was unnecessary because Parcel has its own built-in postcss dependency,
still at v7. (Parcel v2 uses postcss v8, but Parcel v2 is a large upgrade.)

To fix the version conflict, we remove our own Postcss dependency from
package.json, since it's not being used. We also downgrade the
postcss-round-subpixels plugin to 1.2.0, which requires postcss v7.