fix(sequelize): enable, but not enforce ssl for production (#2217)
A relevant snippet of Heroku's support response below:

```
The issue is that even though our documents do state that SSL is required for Postgres connections this is not 100% true for hobby tier databases. We are slowly rolling this out so it is true for ALL databases but some older hobby tier databases do not require SSL when connecting from our dynos (it's still required for external connections).

What is happening here is that your staging and production apps are using a hobby tier database host which doesn't enforce SSL from a dyno so you are not seeing this issue on those apps. When it comes to review apps weather you see this issue or not depends on the host the new Postgres add-on is placed. The newer hosts you will see this issue, older hosts you will not.

It should be noted that if you ever upgraded your staging/production apps to professional tier database (standard-0 and above) you would also see this issue.

In general, you should always ensure that your connection is using SSL regardless of the database plan you use, so adding ssl: true is the right thing to do here. However, this brings up the second error you see ERROR: self-signed certificate. This happens because currently, all our data services by default use a self-signed certificate and the pg NodeJS module settings require the SSL certificate of the Postgres host to be signed by a trusted root CA certificate. We are aware that this isn't ideal and do want to address this in the future.

To work around this error you will need to add rejectUnauthorized: false to your dialectOptions. See more details on this here: https://github.com/brianc/node-postgres/issues/2009
```