test(react-dnd): refactor test implementation
This commit began as an attempt to remove the `React.createFactory() will
be deprecated` warnings that would occur during unit tests. The warning came
from the `recompose` package, which is no longer being updated. This meant
that any solution would require replacing the `recompose` package, either
with a similar one, or with different functionality.

We only use this package to set up a mock version of the react-dnd library
for testing.

So, instead of replacing `recompose`, I refocused my attention on writing
the tests differently. The solution (based on react-dnd's documentation)
would be to wrap components under test, rather than to stub out the library
entirely.

Looking at react-testing-library docs, we also see some patterns for
testing components that are wrapped with providers that we could adopt to
improve and streamline our own codebase.

This commit is targeted only for react-dnd implementation, but also
provides the initial step in putting all of our provider wrappers into
one single method (instead of `renderWithIntl` or `renderWithRedux`
separately).