refactor(i18n): remove use_local configuration (#2248)
This is driven by the eventual goal to remove the staging environment
configuration. (Staging should be as close to Production as possible,
just on a different instance.)

To reach that goal, any staging-specific setting in `config` must be
achieved through some other means, since `config/staging.js` will
no longer be read. For internationalization (i18n) this means that
the `l10n.use_local` configuration setting needs to be applied
in another way.

The simplest way to achieve this is to change the decision making of
whether or not to use local translation files based on the presence
of the `TRANSIFEX_API_TOKEN` alone. If present, it uses the Transifex
API. If not present, it reads local translation files.

Our production server, which reads local translation files, will never
need the Transifex API, so we can remove the token from its env vars.
In staging, we add the Transifex token to allow live translations.
However, we can also unset the variable to achieve the same behavior
as production (and this can be done without pushing configuration
changes to the repository.)

This will also be true for all local instances. By default, without
a Transifex API token, local instances will always read from local
translation files. Only local instances with tokens will read from
the API. Again, to go back to reading local files, just remove the
API token (no need for a second setting to toggle.)