fix(locales): use mkdirp to create directories
Previously, the `bin/download_translations.js` script used the
`fs.stat` utility (built in to Node) to check the existence of
a directory, and if it didn't exist, to create the directory.

But it doesn't work (or did it used to work, but not anymore?)

According to Node.js docs:
(https://nodejs.org/api/fs.html#fs_fs_stat_path_options_callback)

> Using fs.stat() to check for the existence of a file ...
> is not recommended.

Anyway, the more common way that Node developers will accomplish
this same task is to use the `mkdirp` module. We already use it
in `lib/svg-sprite.js`, although we don't specify the dependency
directly in `package.json`. (It's a common enough package that
it was already included by another dependency.)

In this PR, we explicitly include `mkdirp` as a dependency, and
use it to create new folders, when necessary, as translations
are downloaded.