feat: make auth0 domain fully configurable #2330
Previously, we allowed for an `AUTH0_DOMAIN` env configuration setting,
but weren’t fully committed to it: parts of the codebase still
hard-referenced `streetmix.auth0.com`.

Now, we properly dribble the `AUTH0_DOMAIN` env var through the full
stack, setting the app up for alternate Auth0 tenants.

This also includes a very minor fix to the URL for our `jwksUri` JSON
Web Key Set, which was set to the HTTP version of the URL instead of
HTTPS. I updated that while adding the string concatenation.

Closes #2330