fix(db): let profile_image_url migration roll back
Rolling back a column length change fails when it would result in
destroying data. This makes sense, but we can’t opt out of that
protection even though we’re willing to accept it in this context.

This commit thus replaces the simple migration reversal with a more
complex one, wrapped in a transaction, which first lops off all
`Users.profile_image_url` values greater than 255, then changes the
column length.

It’s been tested and works, though it drops into raw SQL to do so, and
so can probably be improved.

I also snuck in an npm run shortcut for rolling back a migration.

Affects #2335