fix(ShareMenu): broken CopyIcon
@primer/octicons-react@13.0.0 and later removed the CopyIcon component,
appearing to have renamed it to DuplicateIcon instead. This fix updates
the import so that it is now importing the correct icon. When it was
broken, trying to open the ShareMenu caused the menu component to crash.

This also updates @primer/octicons-react to the latest version as of this
commit and also does a slight visual tweak to keep the 16px icon more in
line with the application's current base font size (13px).