chore(deps): bump babel/* packages to 7.16.x except preset-env, see note
The @babel/preset-env package starting at v7.13.9 and above will throw a
"Invalid Version: undefined" error on build time, as a result of a conflict
with parcel-bundler v1. See discussion here:

https://github.com/parcel-bundler/parcel/issues/5943

There is no fix in Parcel v1, so @babel/preset-env should be pegged here
until Parcel is upgraded to v2 or another solution is found.