Bump remainder of dependencies that can be easily updated (#2375)
* chore(deps): bump sass

Note: I tried also to bump postcss and related plugins to support postcss v8,
but postcss is locked to v7 by parcel-bundler.

* chore(browserslist): drop IE11 explicitly in favor of "not dead" browsers

* chore(deps): express and various server-side packages

* chore(deps): bump core-js, regenerator-runtime

* chore(deps): remove unused dateformat

* chore(deps): bump some front-end auth deps

* chore(deps): bump 3rd-party libs

Note: could not update cloudinary@1.28.1 because it throws a
"ReferenceError: setImmediate is not defined" error in jest

* chore(deps): bump postgres deps

* chore(deps): bump misc low-impact package updates

* chore(deps): bump various low-impact react deps

* chore(deps): bump intl deps

* chore(deps): bump redux toolkit

* chore(deps): bump react-dnd, requires downgrading react-dnd-multi-backend

* chore(deps): bump react

Somehow we still need to also add our own version of postcss@^8 for
stylelint. Note that it's not used by parcel or other postcss plugins
which is based on parcel-bundler's postcss@^7.