test: remove act()
removes a warning in some tests that "this environment is not configured to support act()"
this might be a deprecation for React 19:
https://react.dev/blog/2024/04/25/react-19-upgrade-guide#removed-react-dom-test-utils

for consistency, act() is removed in every test, even if it did not throw the warning, if
it doesn't affect the actual test output. if act() is required again, put it back later.