test(StreetNameplateContainer): handle /api/v1/users/ call
for some reason, this endpoint was being called during vitest, and msw warned that it wasn't
being handled. not sure why it's being called, it isn't called in any other component or
child component. the likely theory is that it's a side effect of dispatching a new street
name to the redux store. but this also doesn't happen in the client, so i found it hard to
trace.

workaround is just to put the path in msw's handlers and return a 401, as it should do if
this path is called in real life.