refactor(Gallery): fix selector warning in vitest
"Selector unknown returned a different result when called with the
same parameters. This can lead to unnecessary rerenders."
Message appears when running the "closes on shield click" test for
<Gallery />. This was caused by the selector in a child component
<GalleryContents /> returning a new empty array whenever the store's
gallery.streets property didn't exist.

In reality the property always exists and is initialized as an
empty array, with streets in the gallery cached in memory when it's
hidden. The only time it's _not_ initialized is in test, because it's
given an initialState where the streets property wasn't present.

Fixed by:
- Updating the test to use a more complete initialState
- Removing the empty array fallback on <GalleryContents />. There
  shouldn't need to be one, and if it's missing, that is an error.

Also included: some refactoring of adjacent components.