Housekeeping (#3396)
* fix(StreetNameplateContainer): port to ts + tweaks

- make it possible to unname a street (handles empty string)

* fix(api/users): properly return errors

* fix(api): return logic for all users

* test(StreetNameplateContainer): handle /api/v1/users/ call

for some reason, this endpoint was being called during vitest, and msw warned that it wasn't
being handled. not sure why it's being called, it isn't called in any other component or
child component. the likely theory is that it's a side effect of dispatching a new street
name to the redux store. but this also doesn't happen in the client, so i found it hard to
trace.

workaround is just to put the path in msw's handlers and return a 401, as it should do if
this path is called in real life.

* chore(packages): bump dependencies

* refactor(Gallery): fix selector warning in vitest

"Selector unknown returned a different result when called with the
same parameters. This can lead to unnecessary rerenders."
Message appears when running the "closes on shield click" test for
<Gallery />. This was caused by the selector in a child component
<GalleryContents /> returning a new empty array whenever the store's
gallery.streets property didn't exist.

In reality the property always exists and is initialized as an
empty array, with streets in the gallery cached in memory when it's
hidden. The only time it's _not_ initialized is in test, because it's
given an initialState where the streets property wasn't present.

Fixed by:
- Updating the test to use a more complete initialState
- Removing the empty array fallback on <GalleryContents />. There
  shouldn't need to be one, and if it's missing, that is an error.

Also included: some refactoring of adjacent components.

* refactor(WhatsNewDialog): remove parcel/react-markdown workaround

* refactor(segment/info): partial port to typescript

todo: handle sprite defs, and try to consolidate/simplify the segment lookup logic

* refactor(info): add type definitions for sprites

* test(info): port test to typescript, add test mock

* refactor(sprites): definition file in yaml

* chore(packages): bump dependencies